extract getDependencies method to AbstractElement
