+++ models/Element/AbstractElement.php
@@ -32,6 +32,11 @@ abstract class AbstractElement extends Model\AbstractModel implements ElementInt
     use DirtyIndicatorTrait;
 
     /**
+     * @var Model\Dependency|null
+     */
+    protected $dependencies;
+
+    /**
      * @var int
      */
     protected $__dataVersionTimestamp = null;
@@ -348,7 +353,14 @@ abstract class AbstractElement extends Model\AbstractModel implements ElementInt
     /**
      * @return Model\Dependency
      */
-    abstract public function getDependencies();
+    public function getDependencies()
+    {
+        if (!$this->dependencies) {
+            $this->dependencies = Model\Dependency::getBySourceId($this->getId(), Service::getElementType($this));
+        }
+
+        return $this->dependencies;
+    }
 
     /**
      * @return Model\Schedule\Task[]
@@ -364,5 +376,28 @@ abstract class AbstractElement extends Model\AbstractModel implements ElementInt
     public function getVersions()
     {
         return [];
+    }
+
+    /**
+     * @return array
+     */
+    public function __sleep()
+    {
+        $finalVars = [];
+        $blockedVars = ['dependencies'];
+        $vars = get_object_vars($this);
+        foreach ($vars as $key => $value) {
+            if (!in_array($key, $blockedVars)) {
+                $finalVars[] = $key;
+            }
+        }
+
+        return $finalVars;
+    }
+
+    public function __clone()
+    {
+        parent::__clone();
+        $this->dependencies = null;
     }
 }
